--- test/operator_group.yaml
@@ -1,14 +1,0 @@
-apiVersion: operators.coreos.com/v1
-kind: OperatorGroup
-metadata:
-  annotations:
-    olm.providedAPIs: AerospikeCluster.v1beta1.asdb.aerospike.com
-  name: test-operator-group
-  namespace: test
-spec:
-  targetNamespaces:
-    - aerospike
-    - test
-    - test1
-    - test2
-  upgradeStrategy: Default
