--- test/subscription.yaml
@@ -1,11 +1,0 @@
-apiVersion: operators.coreos.com/v1alpha1
-kind: Subscription
-metadata:
-  name: aerospike-kubernetes-operator
-  namespace: test
-spec:
-  channel: stable
-  installPlanApproval: Automatic
-  name: aerospike-kubernetes-operator
-  source: aerospike-kubernetes-operator
-  sourceNamespace: test
